clk: qcom: Hook up gpu_gx gdsc as a parent of gpu_gdsc
the GPU device on msm8996 requires 2 gdscs to be powered on.
With powerdomains we have a limitation wherein only one
powerdomain can be associated with a given device in DT.
So hook the 2 gdscs together so both are powered on/off when
the gpu driver needs them to.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
(cherry picked from commit 90b9f9283df99755399dd68137a471a07b81ef66)