kernel: configs: set USB_CONFIG_F_FS in distro.config
We initially add it to arch/arm64/defconfig, however USB_CONFIG_F_FS depends on
USB_CONFIGFS which is not set in arch/arm64/defconfig, so doing a savedefconfig
would prune it. Since we have USB_CONFIGFS in kernel/configs/distro.config , it
is better to set USB_CONFIG_F_FS there as well.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>