UBUNTU: [Packaging] Fix update-from-* abi file handling
The update script should always have been getting the ABI
files from the last build. In order to fix this a few odd
looking steps have to be taken.

The problem is that call to "debian/rules clean" will delete
any abi directory named with the current package version. So
the download has to happen after creating the new version in
the changelog.

Before cranky, the update script would copy the ABI files from
the master directory (so the version name was always not equal
to the current package version), then do a clean and updateconfigs.
Since the version was fake, it would survive the clean and then
get staged.
To support the old style calling, the section of downloading the
ABI files has to move after the clean call.

With cranky, however, the changes to the changelog by the update
script will be thrown away and then another clean run is done
which again would dump all the ABI files (since the version got
reset). So if called from cranky, we will not bother to do anything
to the ABI files and let that get handled by the startnewrelease
call (maint-startnewrelease) which needs a different fix as it
currently ignores getting the ABI when a version has a tilde in it.

Ignore: yes

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>