UBUNTU: [Packaging] Support building libperf-jvmti.so
BugLink: https://bugs.launchpad.net/bugs/1761379

Adds default-jdk-headless and java-common as Build-Depends, which will
allow libperf-jvmti.so to be built.

The library will be installed on the linux-tools package the same way other
tools are installed. That allows a user to use the current kernel version
as given by `uname -r` to find the library at
/usr/lib/linux-tools/`uname -r`/libperf-jvmti.so, which will be a symlink
to a version-specific library.

This requires arches and derivatives to opt in with do_tools_perf_jvmti.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>