UBUNTU: [Config] raspi2: Unset CONFIG_MODVERSIONS
BugLink: https://bugs.launchpad.net/bugs/1839213

This has been turned off in the master kernel from the beginning
and was annotated to be turned off on all arches. To work around
a regression introduced by the 2019-08-06 stable update and to
make all kernel builds more look-alike, change it to off.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>