UBUNTU: [Packaging]: disable lz4 when mangling
BugLink: https://bugs.launchpad.net/bugs/1844539

When we copy the config from the master kernel, CONFIG_KERNEL_LZ4 will be
enabled for amd64, i386 and s390x. lz4 is not available on bionic as of this
time, so PPA builds will wait for the build-dependency and not complete.

Fix the local-mangle script so it will disable CONFIG_KERNEL_LZ4 for those
architectures and use the latest used config for each architectures, that is,
CONFIG_KERNEL_GZIP for amd64 and i386, and CONFIG_KERNEL_XZ for s390x.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
[ klebers: applied for linux-azure-5.3. ]
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>